Learning_To_Adapt/SafeRL_WMPC/requirements.txt
# Core dependencies
torch>=1.9.0
numpy>=1.21.0
matplotlib>=3.5.0
pyyaml>=6.0

# Bayesian Optimization
botorch>=0.8.0
gpytorch>=1.8.0

# Reinforcement Learning
stable-baselines3>=2.0.0
gymnasium>=0.29.0

# Optional: for better performance
scipy>=1.7.0
scikit-learn>=1.0.0

# Development and testing
pytest>=6.0.0
pytest-cov>=3.0.0

# Logging and utilities
tqdm>=4.62.0 